+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 3.7.4
+
+- Added `showAssetsLabels`, `showGeofencesLabels`, `showPoisLabels`, and `showAssetsTrails` boolean fields to `Locator` and `LocatorInput`.
+- `enableSidebar` and `showAssetsTrails` default to `true` in `LocatorInput`.
+
 ## 3.7.3
 - Refactor `CaclInput`.